# public/swagger/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bütçe API",
        "version": "1.0"
    },
    "paths": {
...
        "/paymenttypes/findAllByPaging": {
            "get": {
                "tags": [
                    "Payment Types"
                ],
                "summary": "Find all payment types by pagination",
                "operationId": "paymentTypes_findAllByPaging",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_perPage",
                        "in": "query",
                        "description": "Per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_sort",
                        "in": "query",
                        "description": "Order items",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentTypeFindAllByPageResultVM"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No result found"
                    }
                }
            }
        },
        "/paymenttypes/findOneById/{paymentTypeId}": {
            "get": {
                "tags": [
                    "Payment Types"
                ],
                "summary": "Find item data",
                "operationId": "paymentTypes_findOneById",
                "parameters": [
                    {
                        "name": "paymentTypeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentTypeFindOneByIdResultVM"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/paymenttypes/create": {
            "post": {
                "tags": [
                    "Payment Types"
                ],
                "summary": "Create a new expense type",
                "operationId": "paymentTypes_create",
                "requestBody": {
                    "description": "Create new expense type",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentTypeSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
        "/paymenttypes/update/{paymentTypeId}": {
            "put": {
                "tags": [
                    "Payment Types"
                ],
                "summary": "Update expense type",
                "operationId": "paymentTypes_update",
                "parameters": [
                    {
                        "name": "paymentTypeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update customer",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentTypeSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request, returns to validation errors"
                    }
                }
            }
        },
        "/paymenttypes/delete/{paymentTypeId}": {
            "delete": {
                "tags": [
                    "Payment Types"
                ],
                "summary": "Delete expense type",
                "operationId": "paymentTypes_delete",
                "parameters": [
                    {
                        "name": "paymentTypeId",
                        "in": "path",
                        "description": "PaymentType uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaymentTypeFindAllByPageResultVM": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PaymentTypeListItem"
                        }
                    },
                    "page": {
                        "type": "integer"
                    },
                    "perPage": {
                        "type": "integer"
                    },
                    "totalPages": {
                        "type": "integer"
                    },
                    "totalItems": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PaymentTypeFindOneById": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "paymentTypeName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PaymentTypeFindOneByIdResultVM": {
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/PaymentTypeFindOneById"
                    }
                },
                "type": "object"
            },
            "PaymentTypeListItem": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "paymentTypeName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "PaymentTypeSave": {
                "properties": {
                    "paymentTypeId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "paymentTypeName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}